requirements.txt
Django>=4.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9